s05/requirements.txt
asgiref==3.8.1
django==5.1.1
psycopg2-binary==2.9.9
sqlparse==0.5.1
typing_extensions==4.12.2
